✨ feat(format): Modify formatJSON function for i18n workflow
This commit introduces changes to the format.mjs file. The formatJSON
function has been modified to handle the i18n workflow. The changes
involve setting values in the rawData object based on selectors and
checking for the existence of rawData before executing certain code.

Changes made:
- Modifications to the formatJSON function for i18n workflow

The purpose of these changes is to improve the functionality and
performance of the i18n workflow in the formatJSON function.